+++ src/controls.js
@@ -188,7 +188,11 @@ _V_.LoadingSpinner = _V_.Component.extend({
 
     player.addEvent("seeking", _V_.proxy(this, this.show));
     player.addEvent("error", _V_.proxy(this, this.show));
-    player.addEvent("stalled", _V_.proxy(this, this.show));
+
+    // Not showing spinner on stalled any more. Browsers may stall and then not trigger any events that would remove the spinner.
+    // Checked in Chrome 16 and Safari 5.1.2. http://help.videojs.com/discussions/problems/883-why-is-the-download-progress-showing
+    // player.addEvent("stalled", _V_.proxy(this, this.show));
+
     player.addEvent("waiting", _V_.proxy(this, this.show));
   },