+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Stopped spinner from showing on 'stalled' events since browsers sometimes don't show that they've recovered.
 * Fixed CDN Version which was breaking dev.html
 * Made full-window mode more independent
 * Added rakefile for release generation